+++ packages/eslint-plugin-conventions/src/index.ts
@@ -56,7 +56,7 @@ withCanonicalDocsUrls('plugin-conventions', rules);
 export const plugin = {
   meta: {
     name: 'eslint-plugin-conventions',
-    version: '6.0.7',
+    version: '6.0.8',
   },
   rules,
 } satisfies TSESLint.FlatConfig.Plugin;