+++ packages/eslint-plugin-conventions/src/rules/conventions/no-commented-code.ts
@@ -314,7 +314,7 @@ export const noCommentedCode = createRule<RuleOptions, MessageIds>({
 
     const filename = context.filename;
     const isTestFile =
-      ignoreInTests && /\.(test|spec)\.(ts|tsx|js|jsx)$/.test(filename);
+      ignoreInTests && /\.(test|spec)\.[cm]?[jt]sx?$/.test(filename);
 
     if (isTestFile) {
       return {};