+++ packages/eslint-plugin-react-a11y/src/rules/no-keyboard-inaccessible-elements.ts
@@ -147,7 +147,7 @@ export const noKeyboardInaccessibleElements = createRule<
 
     const filename = context.filename;
     const isTestFile =
-      ignoreInTests && /\.(test|spec)\.(ts|tsx|js|jsx)$/.test(filename);
+      ignoreInTests && /\.(test|spec)\.[cm]?[jt]sx?$/.test(filename);
 
     if (isTestFile) {
       return {};